handson/requirements.txt
gradio
google-cloud-discoveryengine
google-api-core
